\name{downloadOpenMLDataSetDescription}
\alias{downloadOpenMLDataSetDescription}

\title{
  Downloading an openML task from the repository.
}
\description{
 This function downloads an openML task from the respository saving is
  information into a local XML file and returning the task information
  as a \code{OpenMLTask} object.
}

\usage{
downloadOpenMLDataSetDescription(id, file)
}

\arguments{
  \item{id}{

  }
  \item{file}{

  }
}

\details{

}

\references{ 

  \url{http://www.dcc.fc.up.pt/~ltorgo/openML}  
}

\author{ Luis Torgo \email{ltorgo@dcc.fc.up.pt}, Bernd Bischl \email{}
  and Paula Branco \email{paobranco@gmail.com}
}

\seealso{
  \code{\link{downloadOpenMLDataSetDescription}}
}

\examples{

}

\keyword{models}

