\name{getDataCharacteristics}
\alias{getDataCharacteristics}
\title{Get basic data characteristics.}
\usage{
  getDataCharacteristics()
}
\value{
  [\code{data.frame}]. A \code{data.frame} containing the
  data characteristics.
}
\description{
  Retrieve the following "non-meta" data characteristics:
  "NumberOfFeatures", "NumberOfInstances",
  "NumberOfClasses", "NumberOfInstancesWithMissingValues",
  "NumberOfMissingValues", "NumberOfNumericFeatures",
  "NumberOfSymbolicFeatures"
}

