\name{getDataChars}
\alias{getDataChars}
\title{getDataChars}
\usage{
  getDataChars()
}
\value{
  [\code{data.frame}]. A \code{data.frame} containing the
  data qualities.
}
\description{
  Retrieve the following "non-meta" data characteristics:
  "NumberOfFeatures", "NumberOfInstances",
  "NumberOfClasses", "NumberOfInstancesWithMissingValues",
  "NumberOfMissingValues", "NumberOfNumericFeatures",
  "NumberOfSymbolicFeatures", "PublicationDate"
}

