\name{OptionalOpenMLDataSetDescription}
\alias{OptionalOpenMLDataSetDescription}
\alias{OptionalOpenMLDataSetDescription-class}
\title{OptionalOpenMLDataSetDescription}
\description{
Either an object of class
\code{\link{OpenMLDataSetDescription}} or \code{NULL}.
}
\seealso{
\code{\link{OpenMLDataSetDescription}},
\code{\link{OpenMLTask}}
}

